backend/requirements.txt
# FastAPI Backend Dependencies
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlmodel==0.0.22
groq==0.13.0
python-dotenv==1.0.1
openai>=1.0.0

# PostgreSQL Support
psycopg2-binary==2.9.10
asyncpg==0.30.0
sqlalchemy[asyncio]==2.0.36

# Google APIs
google-api-python-client==2.158.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1

# Email Service
aiosmtplib==3.0.2

# HTTP Client for Slack
httpx==0.28.1

# Date/Time Utilities
python-dateutil==2.9.0

# Twilio and JWT dependencies
twilio==9.4.0
python-jose[cryptography]==3.3.00
reportlab==4.1.0
fastmcp==0.1.0
